add utils.py and cleanup composition_features.py
