+++ matminer/descriptors/composition_features.py
@@ -3,16 +3,14 @@ from __future__ import division, unicode_literals, print_function
 import itertools
 import re
 from collections import namedtuple
-from functools import reduce
 
 import numpy as np
-
-from pymatgen import Element, Composition, MPRester
+from matminer.descriptors.base import AbstractFeaturizer
+from matminer.descriptors.data import magpie_data
+from matminer.descriptors.utils import get_holder_mean
+from pymatgen import Element, Composition
 from pymatgen.core.periodic_table import get_el_sp
 from pymatgen.core.units import Unit
-
-from matminer.descriptors.base import AbstractFeaturizer
-from matminer.descriptors.data import magpie_data, cohesive_energy_data
 
 __author__ = 'Jimin Chen, Logan Ward, Saurabh Bajaj, Anubhav jain, Kiran Mathew'
 
@@ -282,82 +280,6 @@ def get_pymatgen_descriptor(composition, property_name):
             eldata.append(property_value)
 
     return eldata
-
-
-def get_cohesive_energy(comp):
-    """
-    Get cohesive energy of compound by subtracting elemental cohesive energies from the formation energy of the compund.
-    Elemental cohesive energies are taken from http://www.knowledgedoor.com/2/elements_handbook/cohesive_energy.html.
-    Most of them are taken from "Charles Kittel: Introduction to Solid State Physics, 8th edition. Hoboken, NJ:
-    John Wiley & Sons, Inc, 2005, p. 50."
-
-    Args:
-        comp: (str) compound composition, eg: "NaCl"
-
-    Returns: (float) cohesive energy of compound
-
-    """
-    el_amt_dict = Composition(comp).get_el_amt_dict()
-
-    # Get formation energy of most stable structure from MP
-    struct_lst = MPRester().get_data(comp)
-    if len(struct_lst) > 0:
-        struct_lst = sorted(struct_lst, key=lambda e: e['energy_per_atom'])
-        most_stable_entry = struct_lst[0]
-        formation_energy = most_stable_entry['formation_energy_per_atom']
-    else:
-        raise ValueError('No structure found in MP for {}'.format(comp))
-
-    # Subtract elemental cohesive energies from formation energy
-    cohesive_energy = formation_energy
-    for el in el_amt_dict:
-        cohesive_energy -= el_amt_dict[el] * cohesive_energy_data[el]
-
-    return cohesive_energy
-
-
-def band_center(comp):
-    """
-    Estimate absolution position of band center using geometric mean of electronegativity
-    Ref: Butler, M. a. & Ginley, D. S. Prediction of Flatband Potentials at Semiconductor-Electrolyte Interfaces from
-    Atomic Electronegativities. J. Electrochem. Soc. 125, 228 (1978).
-
-    Args:
-        comp: (Composition)
-
-    Returns: (float) band center
-
-    """
-    prod = 1.0
-    for el, amt in comp.get_el_amt_dict().iteritems():
-        prod = prod * (Element(el).X ** amt)
-
-    return -prod ** (1 / sum(comp.get_el_amt_dict().values()))
-
-
-def get_holder_mean(data_lst, power):
-    """
-    Get Holder mean
-
-    Args:
-        data_lst: (list/array) of values
-        power: (int/float) non-zero real number
-
-    Returns: Holder mean
-
-    """
-    # Function for calculating Geometric mean
-    geomean = lambda n: reduce(lambda x, y: x * y, n) ** (1.0 / len(n))
-
-    # If power=0, return geometric mean
-    if power == 0:
-        return geomean(data_lst)
-
-    else:
-        total = 0.0
-        for value in data_lst:
-            total += value ** power
-        return (total / len(data_lst)) ** (1 / float(power))
 
 
 def get_composition_oxidation_state(formula):
